src/test/java/com/betrue/kkanbufs_be/service/AuthServiceTest.java
package com.betrue.kkanbufs_be.service;


import com.betrue.kkanbufs_be.domain.User;
import com.betrue.kkanbufs_be.exception.AlreadyExistsEmailException;
import com.betrue.kkanbufs_be.repository.UserRepository;
import com.betrue.kkanbufs_be.request.Signup;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import static org.junit.jupiter.api.Assertions.*;
import static org.junit.jupiter.api.Assertions.assertThrows;

@SpringBootTest
class AuthServiceTest {
    @Autowired
    private UserRepository userRepository;

    @Autowired
    private AuthService authService;

    @AfterEach
    void clean() {
        userRepository.deleteAll();
    }

    @Test
    @DisplayName("회원가입 성공")
    void test1() {
        // given
        Signup signup = Signup.builder()
                .loginId("userid")
                .password("1234")
                .name("김동혁")
                .build();

        // when
        authService.signup(signup);

        // then
       assertEquals(1, userRepository.count());

        User user = userRepository.findAll().iterator().next();
        assertEquals("userid", user.getLoginId());
        assertEquals("1234", user.getPassword());
        assertEquals("김동혁", user.getName());
    }

    @Test
    @DisplayName("회원가입시 중복된 이메일")
    void test2() {
        // given
        User user = User.builder()
                .loginId("hodolman88@gmail.com")
                .password("1234")
                .name("짱돌맨")
                .build();
        userRepository.save(user);

        Signup signup = Signup.builder()
                .loginId("hodolman88@gmail.com")
                .password("1234")
                .name("호돌맨")
                .build();

        // expected
        assertThrows(AlreadyExistsEmailException.class, () -> authService.signup(signup));
    }
}